requirements.txt
text2emotion
ibm-watson
transformers==3.0.0
opencv-python
imutils==0.5.4
matplotlib
pandas
Pillow
plotly
streamlit
pyjokes
wikipedia
watchdog
quotes-generator
wolframalpha
h5py==2.10.0
numpy
SpeechRecognition
newsapi
